+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 # Version 3
 
+## [Unreleased]
+### Changed
+- [git diff](/../../compare/3.0.1...master)
+
 ## [3.0] - 2016-03-15
 ### Added
 - services